[Client] Timer 패킷 추가 및 Recv 테스트 완료, 타이머 int로 바꿈, Progress Bar 그리기 수정 필요

+++ NGP_project/Client/GameObject.h
@@ -67,7 +67,7 @@ public:
 public:
 	void ResetCurFrame();
 
-	float _timer{};	// 총알, 폭탄 등에 사용
+	int _timer{};	// 총알, 폭탄 등에 사용
 	Status _status{0, 10};
 protected:
 	HBITMAP _bitmap{};
